+++ src/common/utxobased/info/all.ts
@@ -18,7 +18,6 @@ import { info as litecoin } from './litecoin'
 import { info as pivx } from './pivx'
 import { info as qtum } from './qtum'
 import { info as ravencoin } from './ravencoin'
-import { info as smartcash } from './smartcash'
 import { info as ufo } from './ufo'
 import { info as vertcoin } from './vertcoin'
 import { info as zcoin } from './zcoin'
@@ -43,7 +42,6 @@ export { info as litecoin } from './litecoin'
 export { info as pivx } from './pivx'
 export { info as qtum } from './qtum'
 export { info as ravencoin } from './ravencoin'
-export { info as smartcash } from './smartcash'
 export { info as ufo } from './ufo'
 export { info as vertcoin } from './vertcoin'
 export { info as zcoin } from './zcoin'
@@ -69,7 +67,6 @@ export const all = [
   pivx,
   qtum,
   ravencoin,
-  smartcash,
   ufo,
   vertcoin,
   zcoin