+++ src/common/utxobased/keymanager/declarations/bs58smartcheck.d.ts
@@ -1,5 +1,0 @@
-declare module 'bs58smartcheck' {
-  export const encode: (buffer: Buffer | number[] | Uint8Array) => string
-  export const decodeUnsafe: (string: string) => Buffer | undefined
-  export const decode: (string: string) => Buffer
-}